+++ matlab/src/matlab/featherwrite.m
